refac: use folder to distinguish resource
