+++ src/hooks/useServerList.ts
@@ -1,5 +1,5 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
-import { getServerList } from '../api/getServerList';
+import { getServerList } from '../api/server/getServerList';
 
 export const UseServerListQueryKey = "useServerListQueryKey";
